requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
tushare==2.0.0
talib-binary==0.4.24 